requirements.txt
django==1.4.5
PIL
South
pygments
GitPython
Django403Handler
markdown
django-pipeline==1.3.4
MySQL-python
python-ldap
django-celery-with-redis